Add home page templating
Using Jinja template engine, add templating for
the first home page of SWESphere.
Implement a base template so future pages have
the same features such as header and footer.
We want this to be consistent across all pages
of the site, so implementing a base template is
how we can ensure this, as the site evolves.