Bug Fix - Changing Username to one that exists
If a user tried to change their username to that
of another user in the system, there would be an
issue because each username in the system must be
unique per the User schema. This bugfix validates
the changed username by ensuring that the new
username isn't taken by another user.